refactor: improve error handling, naming consistency, and test reusability
- Updated error handling to use `errors.Is` and `errors.As` for improved readability and accuracy.
- Standardized struct naming (`streamValue` → `StreamValue`, `sortedSet` → `SortedSet`, etc.) for consistency.
- Extracted common test configuration setup into `defaultTestConfig` to improve test reusability.
- Replaced redundant inline shutdown logic with `waitForServerStop` utility function for cleaner test cases.
- Removed unused `String` function and redundant `ErrValueNotFloat` error.
- Ensured proper usage of constants in stream utilities and memory management.

+++ internal/rdb/loader.go
@@ -231,7 +231,7 @@ func (l *loader) readValueMarker(marker byte) error {
 }
 
 func validatedExpiryMillis(expiresAtMillis uint64) (int64, error) {
-	if expiresAtMillis > uint64(^uint64(0)>>1) {
+	if expiresAtMillis > ^uint64(0)>>1 {
 		return 0, fmt.Errorf("rdb: expiry milliseconds overflow: %d", expiresAtMillis)
 	}
